DFA

Description

The DFA holds a list of DFAStates, each of which hold links to other states on reading token, similarly to how a Turing Machine DFA works. These states also hold information about the productions they correspond to, but only by index in the DFAProduction list. Various other HashMaps are also stored within the DFA to keep the structure fully self referencial and standalone.

Documentation

data DFA

The deterministic finite automata

Constructors

DFA

Fields

_dfaStates :: [DFAState]

List of states, identified by their position in the list

_dfaProductions :: [DFAProduction]

All productions simplified from Grammar

_dfaTokenMap :: TokenMap

Map from token name to token pattern

_dfaPrecMap :: PrecMap

Map from terminal to Prec

_dfaFollowMap :: FollowMap

Map from non terminal to list of terminals

data Prec

Combination of Associativity and a precedence level index

Constructors

Prec Associativity Int

type PrecMap = HashMap TokenDef Prec

Map from terminal to Prec

type TokenMap = HashMap String TokenDef

Map from terminal name to TokenDef

type FollowMap = HashMap String [TokenDef]

Map from non terminal to list of terminals

data DFAProduction

Simplified Production from RuleProduction, now storing its own non terminal name and rid of RuleTokenModifier

Constructors

DFAProduction

Fields

_dfaProductionName :: String

_dfaProductionTokens :: [RuleTokenType]

_dfaProductionResult :: String

_dfaProductionPrec :: Maybe Prec

data DFAAction

Action to be taken on reading a token in the DFA

Constructors

DFAShift Int
Shift current token and change to state by index

DFAReduce Int
Reduce by production from index

DFAFinish
Finish reading tokens, language accepted.

data DFAState

DFA state includes production positions and list of actions.

Constructors

DFAState

Fields

productions :: [(Int, Int)]

List of pairs of production index and position within production

actions :: [(RuleTokenType, DFAAction)]

List of pairs of token to read followed by action to take

dfaTokenMap :: Lens' DFA TokenMap

dfaStates :: Lens' DFA [DFAState]

dfaProductions :: Lens' DFA [DFAProduction]

dfaPrecMap :: Lens' DFA PrecMap

dfaFollowMap :: Lens' DFA FollowMap

dfaProductionTokens :: Lens' DFAProduction [RuleTokenType]

dfaProductionResult :: Lens' DFAProduction String

dfaProductionPrec :: Lens' DFAProduction (Maybe Prec)

dfaProductionName :: Lens' DFAProduction String