Grammar

Description

This module holds the Grammar data structure definition, used for generating the DFA structure, as well as various various getters for that structure. It also holds the ScannerSpec structure, which is an abstraction of Scanner used for generating the Scanner instance code.

Documentation

data Grammar

Grammar interpreted from the gmr file, contains all information needed to generate a DFA

Constructors

Grammar

Fields

tokenDefs :: [TokenDef]

All tokens definitions for a grammar

precLevels :: [PrecLevel]

Precedence definitions for tokens

rules :: [Rule]

Ordered rule base

data TokenDef

Direct output from parser of token definitions, to be translated to a map.

Constructors

TokenDef

Fields

tokenName :: String

Name used in rules

tokenPattern :: String

Haskell pattern to match in generated code

data Associativity

Parameterless data type for associativity

Constructors

LeftAssoc

RightAssoc

NonAssoc

data PrecLevel

Precedence level with list of tokens and associativity for those tokens

Constructors

PrecLevel

Fields

levelAssociativity :: Associativity

levelTokens :: [String]

tokens as list of TokenDef names

data Rule

Grammar Rule as a list of productions for a non terminal

Constructors

Rule

Fields

ruleName :: String

Name of non terminal

ruleProductions :: [RuleProduction]

data RuleTokenType

Tokens used indirectly in RuleProduction, as either terminals or non terminals

Constructors

RuleNonTerminal String

RuleTerminal String

data RuleTokenModifier

Modifiers for RuleToken, eventually replaced with extra rules

Constructors

RuleTokenModifierNormal
Default rule behaviour, ignored on pre-processing

RuleTokenModifierSome (Maybe RuleTokenType)
One or more matching for a token, with optional separator token

RuleTokenModifierMany (Maybe RuleTokenType)
Zero or more matching for a token, with optional separator token

RuleTokenModifierOptional
Zero or one matching for a token

data RuleToken

Token used within RuleProduction, holding a RuleTokenType and associated RuleTokenModifier

Constructors

RuleToken

Fields

tokenType :: RuleTokenType

tokenModifier :: RuleTokenModifier

getTokenTypeStr :: RuleTokenType -> String

Name getter for both types of RuleTokenType

getTokenStr :: RuleToken -> String

Name getter for RuleToken wrapper

nonTerminalToken :: String -> RuleToken

Creates a RuleToken non terminal from its name

terminalToken :: String -> RuleToken

Creates a RuleToken terminal from a TokenDef name

ruleToken :: RuleTokenType -> RuleToken

Wraps a RuleTokenType with the default RuleTokenModifier

data RuleProduction

Grammar production, describing a rule of the form A :: a B c { haskellCode }

Constructors

RuleProduction

Fields

productionTokens :: [RuleToken]

Tokens of the production

productionResult :: String

Resulting haskell code for reducing this production

productionPrecToken :: Maybe String

Optional production precedence and associativity override

data ScannerSpec

Copy of the scanner object but all fields are an instance of Show (specifically changing the ParserMap to a String)

Constructors

ScannerSpec

Fields

specSeparateCasedIdentifiers :: Bool

specIgnoreWhitespace :: Bool

specIgnoreComments :: Bool

specOperators :: [String]

specKeywords :: [String]

specBlockComment :: Maybe (String, String)

specLineComment :: Maybe String

specParserMap :: Maybe String

scannerSpec :: ScannerSpec

Default definition for a ScannerSpec, for easy modification via record updating

addScannerSpecTokens :: ScannerSpec -> Grammar -> Grammar

Takes a ScannerSpec and Grammar and adds default and appropriate TokenDefs. Adds defaults for all Token constructors in Scanner, generating those for operators and keywords.