ParserCodeGenerator

Description

Code generator for the parser, all validity checks have already passed at this point, thus the single function in this module cannot fail

Documentation

generateCode

Arguments

:: String
The generated modules name

-> Maybe String
Optional additional exports

-> String
Initial haskell code in output

-> ScannerSpec
ScannerSpec as defined in Grammar

-> DFA
DFA as defined in DFA

-> String
Outputted parser haskell code

Takes Scanner and DFA information and output the parser code