ParserGenerator

Description

This module takes a .gmr file (or the contents of such a file in String form) and outputs the code for a Haskell parser defined by that file. This can be called in an Either String String monad, where Right indicates success and gives the output code, and Left indicates an error, and gives a message. Normally however, this will be called using the IO function, handling file input and output itself. The specification of the .gmr file can be found here: https://github.com/samuelWilliams99/haskell_parser_generator#gmr-format

Documentation

runParserGenerator :: String -> IO ()

Takes the path to an input .gmr file, ensures it is correct, then generates the associated .hs file containing the Parser. On failure, the IO will print to console containing ther error, and stop.

generateParser

Arguments

:: String
The input gmr definition

-> String
The module name for the outputted file

-> (Maybe String -> Maybe String)
A map function applied to the exports, in case functions defined in %precode need to be exported

-> Either String String
The output code or error, Right indicates success, Left indicates failure

Generates a parser without the use of the IO monad, for when you wish to handle IO yourself

pathToModule :: String -> String

Converts a full path to a module name, for example: mypathto/parser.hs -> Parser

parserRequirements :: String

ParserRequirements module as a string