ShiftReduce

Description

Uses the Shift-Reduce algorithm to build a DFA from the DFA module. Shift-Reduce conflicts are automatically resolved if sufficient precedence data has been supplied, otherwise an error is thrown. Similarly, Reduce-Reduce conflicts will cause the algorithm to fail.

Documentation

module ShiftReducePreProcess

generateDFA :: Grammar -> Result DFA

Converts a Grammar from Grammar to a DFA, calling functions exported from ShiftReducePreProcess