ShiftReducePreProcess

Description

Handles changing modifiers to simple rules, building the various HashMaps needed, and ensuring the Grammar is semantically well formed.

Documentation

handleModifiers :: Grammar -> Grammar

Converts all token modifiers to extra rules, giving a modifier-less grammar.

makeTokenMap :: [TokenDef] -> Result TokenMap

Builds a HashMap from a list of TokenDefs, erroring on multiple definitions and invalid patterns

makePrecMap :: [PrecLevel] -> TokenMap -> Result PrecMap

Builds a HashMap from TokenDef to Prec using the precedence levels Errors if a token is present in multiple precedence levels

makeProductions :: [Rule] -> [Rule] -> TokenMap -> PrecMap -> Result [DFAProduction]

Converts rules into a list of DFAProductions, checking all tokens within the rule are valid, and assigning the correct precedences.

makeFollowMap :: [DFAProduction] -> TokenMap -> FollowMap

Calculates all the follower tokens for every Non-Terminal in a list of productions. Followers are defined as terminals that can be read directly after a Non-Terminal, taking all DFAProductions into account. More information about how to calculate this can be found here: http://www.cs.nuim.ie/~jpower/Courses/Previous/parsing/node48.html

getNonTerminals :: [DFAProduction] -> [String]

Gets the names of all Non-Terminals from a list of DFAProductions